runtime: rename initRegistry.ts to crossModuleInitRegistry.ts
The name `initRegistry` collided in conversation with each compiled
module's own `__initializeStatic` / `__initializeGlobals` closures.
The registry specifically holds the cross-module lookup table other
modules use to await each other's init, so the more descriptive name
matches what it does.

Pure file rename plus the one consumer (lib/runtime/index.ts) and
the doc references in init-topsort.md. No symbols change names; the
runtime templates import the symbols from `agency-lang/runtime`, not
by file path, so no template changes are needed.

Amp-Thread-ID: https://ampcode.com/threads/T-019e83b5-20ba-771f-9394-12946c4cb1b4
