+++ packages/agency-lang/lib/runtime/crossModuleInitRegistry.ts
